requirements.txt
accelerate==0.5.1
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
datasets==1.18.4
dgl-cu111==0.8.0
dglgo==0.0.1
dill==0.3.4
docutils==0.17.1
dtw==1.4.0
filelock==3.6.0
fonttools==4.30.0
frozenlist==1.3.0
fsspec==2022.2.0
googledrivedownloader==0.4
huggingface-hub==0.4.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.2
isodate==0.6.1
isort==5.10.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
littleutils==0.2.2
MarkupSafe==2.1.0
matplotlib==3.5.1
multidict==6.0.2
multiprocess==0.70.12.2
networkx==2.7.1
numpy==1.22.3
numpydoc==1.2
ogb==1.3.3
outdated==0.2.1
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyarrow==7.0.0
pycodestyle==2.8.0
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
rdflib==6.1.1
regex==2022.3.2
requests==2.27.1
responses==0.18.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.63.0
transformers==4.11.3
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
xxhash==3.0.0
yacs==0.1.8
yarl==1.7.2
zipp==3.7.0
